+++ CHANGES.rst
@@ -11,6 +11,7 @@ upstream of anybox.recipe.odoo, and will change in the future.
 
 anybox.recipe.odoo 1.9.1 (unreleased)
 -------------------------------------
+- github #18: Support develop-dir option from gp.vcsdevelop
 - github #7: template for ``upgrade.py`` was missing for a.r.odoo
 
 anybox.recipe.openerp 1.9.1 (unreleased)